The head test

+++ Server2/controllers/usersControllers.js
@@ -82,6 +82,9 @@ console.log(`email=${req.body.email} phoneNumber=${req.body.phoneNumber}
 async create(req,res,next)
 {
     try {
+        if (!req.session.profile !== 1) { // Check for profile 1 (doctor)
+            return res.status(403).send('Forbidden - Only doctors can create users');
+        }
 
         const result = usersService.create( req.body);
         res.status(201).send(result);
@@ -100,6 +103,9 @@ async create(req,res,next)
 
     async delete(req, res, next) {
         try {
+            if (!req.session.profile !== 1) { // Check for profile 1 (doctor)
+                return res.status(403).send('Forbidden - Only doctors can delete users');
+            }
             const patientId = req.params.userId; // Use req.params.patientId to get the id
             const result = await usersService.delete(patientId);
             res.status(200).send(`${patientId} deleted successfully`); // Use patientId in the response
